Track graffiti ordered index in db

+++ validator/client/validator.go
@@ -91,7 +91,7 @@ type validator struct {
 	graffiti                           []byte
 	voteStats                          voteStats
 	graffitiStruct                     *graffiti.Graffiti
-	graffitiOrderedIndex               uint
+	graffitiOrderedIndex               uint64
 	eipImportBlacklistedPublicKeys     map[[48]byte]bool
 }
